feat(hugoV2): removed parallax libs
There is not any proper way to have gradient with this kind of libs (it never seems to work with safari which seems to consider the gradient blending div as an independent one). Firefox is always behaving as expected though...

Getting things done with a "normal" fixed attribute on background image, which will degrade gracefully, not have any issue with existing browsers and give a nice effect too.